feat: add some files ('message_repository', e.t.c.) for future logic & configure folder structure in project. Add into 'provider' new logic for messages Refs #19
